+++ src/test/java/features/HomeTest.feature
@@ -11,8 +11,14 @@ Feature: Test cases for onliner
   Scenario: User can open Catalog page
 
 
+
+
   Scenario: User can open Realty page
 
 
   Scenario: User can open Baraholka page
+    Given User is on the Home page
+    When User clicks Baraholka link
+    Then User is on Baraholka page
+
